Add more steps, break up files more
