+++ src/getStepStartStates.test.ts
@@ -2,29 +2,34 @@
 
 import { createMockExecutionContext } from "@jupiterone/integration-sdk/testing";
 
+import getStepStartStates from "./getStepStartStates";
+import { invocationConfig } from "./index";
 import {
+  AD_ACCOUNT,
   AD_GROUP_MEMBERS,
   AD_GROUPS,
   AD_USERS,
-  invocationConfig,
-  RM_COMPUTE_NETWORK_RELATIONSHIPS,
+} from "./steps/active-directory";
+import {
   RM_COMPUTE_VIRTUAL_MACHINE_DISKS,
   RM_COMPUTE_VIRTUAL_MACHINE_IMAGES,
   RM_COMPUTE_VIRTUAL_MACHINES,
-  RM_KEYVAULT_VAULTS,
+} from "./steps/resource-manager/compute";
+import { RM_COMPUTE_NETWORK_RELATIONSHIPS } from "./steps/resource-manager/interservice/constants";
+import { RM_KEYVAULT_VAULTS } from "./steps/resource-manager/key-vault";
+import {
   RM_NETWORK_INTERFACES,
   RM_NETWORK_LOAD_BALANCERS,
   RM_NETWORK_PUBLIC_IP_ADDRESSES,
   RM_NETWORK_SECURITY_GROUPS,
   RM_NETWORK_VIRTUAL_NETWORKS,
-  AD_ACCOUNT,
-} from "./index";
+} from "./steps/resource-manager/network";
 import { IntegrationConfig } from "./types";
 
 describe("getStepStartStates", () => {
   test("empty config", () => {
     const context = createMockExecutionContext<IntegrationConfig>();
-    const states = invocationConfig.getStepStartStates!(context);
+    const states = getStepStartStates(context);
     expect(states).toEqual({
       [AD_ACCOUNT]: { disabled: false },
       [AD_GROUPS]: { disabled: true },
@@ -55,7 +60,7 @@ describe("getStepStartStates", () => {
     const context = createMockExecutionContext<IntegrationConfig>({
       instanceConfig: { ingestActiveDirectory: true } as IntegrationConfig,
     });
-    const states = invocationConfig.getStepStartStates!(context);
+    const states = getStepStartStates(context);
     expect(states).toEqual({
       [AD_ACCOUNT]: { disabled: false },
       [AD_GROUPS]: { disabled: false },
@@ -78,7 +83,7 @@ describe("getStepStartStates", () => {
     const context = createMockExecutionContext({
       instanceConfig: { subscriptionId: "1234" } as IntegrationConfig,
     });
-    const states = invocationConfig.getStepStartStates!(context);
+    const states = getStepStartStates(context);
     expect(states).toEqual({
       [AD_ACCOUNT]: { disabled: false },
       [AD_GROUPS]: { disabled: true },
