+++ src/steps/resource-manager/compute/index.ts
@@ -1,15 +1,26 @@
 import { Entity } from "@jupiterone/integration-sdk";
 
 import { createAzureWebLinker } from "../../../azure";
-import { ACCOUNT_ENTITY_TYPE } from "../../../jupiterone";
 import { IntegrationStepContext } from "../../../types";
+import { ACCOUNT_ENTITY_TYPE, AD_ACCOUNT } from "../../active-directory";
 import { ComputeClient } from "./client";
+import {
+  DISK_ENTITY_TYPE,
+  RM_COMPUTE_VIRTUAL_MACHINE_DISKS,
+  RM_COMPUTE_VIRTUAL_MACHINE_IMAGES,
+  RM_COMPUTE_VIRTUAL_MACHINES,
+  VIRTUAL_MACHINE_DISK_RELATIONSHIP_TYPE,
+  VIRTUAL_MACHINE_ENTITY_TYPE,
+  VIRTUAL_MACHINE_IMAGE_ENTITY_TYPE,
+} from "./constants";
 import {
   createDiskEntity,
   createImageEntity,
   createVirtualMachineDiskRelationships,
   createVirtualMachineEntity,
 } from "./converters";
+
+export * from "./constants";
 
 export async function fetchVirtualMachines(
   executionContext: IntegrationStepContext,
@@ -57,3 +68,30 @@ export async function fetchVirtualMachineDisks(
     jobState.addEntity(createDiskEntity(webLinker, data)),
   );
 }
+
+export const computeSteps = [
+  {
+    id: RM_COMPUTE_VIRTUAL_MACHINE_IMAGES,
+    name: "Virtual Machine Disk Images",
+    types: [VIRTUAL_MACHINE_IMAGE_ENTITY_TYPE],
+    dependsOn: [AD_ACCOUNT],
+    executionHandler: fetchVirtualMachineImages,
+  },
+  {
+    id: RM_COMPUTE_VIRTUAL_MACHINE_DISKS,
+    name: "Virtual Machine Disks",
+    types: [DISK_ENTITY_TYPE],
+    dependsOn: [AD_ACCOUNT],
+    executionHandler: fetchVirtualMachineDisks,
+  },
+  {
+    id: RM_COMPUTE_VIRTUAL_MACHINES,
+    name: "Virtual Machines",
+    types: [
+      VIRTUAL_MACHINE_ENTITY_TYPE,
+      VIRTUAL_MACHINE_DISK_RELATIONSHIP_TYPE,
+    ],
+    dependsOn: [AD_ACCOUNT, RM_COMPUTE_VIRTUAL_MACHINE_DISKS],
+    executionHandler: fetchVirtualMachines,
+  },
+];
