+++ src/steps/resource-manager/cosmosdb/constants.ts
@@ -1,0 +1,10 @@
+// Step IDs
+export const RM_COSMOSDB_SQL_DATABASES = "rm-cosmosdb-sql-databases";
+
+// Graph objects
+export const RM_COSMOSDB_ACCOUNT_ENTITY_TYPE = "azure_cosmosdb_account";
+export const RM_COSMOSDB_ACCOUNT_ENTITY_CLASS = ["Account", "Service"];
+
+export const RM_COSMOSDB_SQL_DATABASE_ENTITY_TYPE =
+  "azure_cosmosdb_sql_database";
+export const RM_COSMOSDB_SQL_DATABASE_ENTITY_CLASS = ["Database", "DataStore"];
