+++ src/steps/resource-manager/databases/constants.ts
@@ -1,0 +1,14 @@
+// Step IDs
+export const RM_DATABASE_SQL_DATABASES = "rm-database-sql-databases";
+export const RM_DATABASE_POSTGRESQL_DATABASES =
+  "rm-database-postgresql-databases";
+export const RM_DATABASE_MARIADB_DATABASES = "rm-database-mariadb-databases";
+export const RM_DATABASE_MYSQL_DATABASES = "rm-database-mysql-databases";
+
+// Graph objects
+export const RM_DATABASE_ENTITY_CLASS = ["Database", "DataStore"];
+export const RM_DATABASE_SERVER_ENTITY_CLASS = [
+  "Database",
+  "DataStore",
+  "Host",
+];
