+++ src/steps/resource-manager/databases/mariadb/constants.ts
@@ -1,0 +1,9 @@
+import { generateRelationshipType } from "@jupiterone/integration-sdk";
+
+export const RM_MARIADB_SERVER_ENTITY_TYPE = "azure_mariadb_server";
+export const RM_MARIADB_DATABASE_ENTITY_TYPE = "azure_mariadb_database";
+export const RM_MARIADB_SERVER_DATABASE_RELATIONSHIP_TYPE = generateRelationshipType(
+  "HAS",
+  RM_MARIADB_SERVER_ENTITY_TYPE,
+  RM_MARIADB_DATABASE_ENTITY_TYPE,
+);
