+++ src/steps/resource-manager/databases/postgresql/index.ts
@@ -1,22 +1,19 @@
 import {
   createIntegrationRelationship,
   Entity,
-  generateRelationshipType,
 } from "@jupiterone/integration-sdk";
 
 import { createAzureWebLinker } from "../../../../azure";
-import { ACCOUNT_ENTITY_TYPE } from "../../../../jupiterone";
 import { IntegrationStepContext } from "../../../../types";
+import { ACCOUNT_ENTITY_TYPE } from "../../../active-directory";
 import { createDatabaseEntity, createDbServerEntity } from "../converters";
 import { PostgreSQLClient } from "./client";
-
-export const RM_POSTGRESQL_SERVER_ENTITY_TYPE = "azure_postgresql_server";
-export const RM_POSTGRESQL_DATABASE_ENTITY_TYPE = "azure_postgresql_database";
-export const RM_POSTGRESQL_SERVER_DATABASE_RELATIONSHIP_TYPE = generateRelationshipType(
-  "HAS",
-  RM_POSTGRESQL_SERVER_ENTITY_TYPE,
+import {
   RM_POSTGRESQL_DATABASE_ENTITY_TYPE,
-);
+  RM_POSTGRESQL_SERVER_ENTITY_TYPE,
+} from "./constants";
+
+export * from "./constants";
 
 export async function fetchPostgreSQLDatabases(
   executionContext: IntegrationStepContext,
