+++ src/steps/resource-manager/databases/sql/index.ts
@@ -1,27 +1,24 @@
 import {
   createIntegrationRelationship,
   Entity,
-  generateRelationshipType,
 } from "@jupiterone/integration-sdk";
 
 import { createAzureWebLinker } from "../../../../azure";
-import { ACCOUNT_ENTITY_TYPE } from "../../../../jupiterone";
 import { IntegrationStepContext } from "../../../../types";
+import { ACCOUNT_ENTITY_TYPE } from "../../../active-directory";
 import { createDatabaseEntity, createDbServerEntity } from "../converters";
 import { SQLClient } from "./client";
+import {
+  RM_SQL_DATABASE_ENTITY_TYPE,
+  RM_SQL_SERVER_ENTITY_TYPE,
+} from "./constants";
 import {
   setAuditingStatus,
   setDatabaseEncryption,
   setServerSecurityAlerting,
 } from "./converters";
 
-export const RM_SQL_SERVER_ENTITY_TYPE = "azure_sql_server";
-export const RM_SQL_DATABASE_ENTITY_TYPE = "azure_sql_database";
-export const RM_SQL_SERVER_DATABASE_RELATIONSHIP_TYPE = generateRelationshipType(
-  "HAS",
-  RM_SQL_SERVER_ENTITY_TYPE,
-  RM_SQL_DATABASE_ENTITY_TYPE,
-);
+export * from "./constants";
 
 export async function fetchSQLDatabases(
   executionContext: IntegrationStepContext,
