+++ src/steps/resource-manager/interservice/constants.ts
@@ -1,0 +1,9 @@
+// Step IDs
+export const RM_COMPUTE_NETWORK_RELATIONSHIPS =
+  "rm-compute-network-relationships";
+
+// Graph objects
+export const VIRTUAL_MACHINE_NIC_RELATIONSHIP_TYPE = "azure_vm_uses_nic";
+export const VIRTUAL_MACHINE_PUBLIC_IP_ADDRESS_RELATIONSHIP_TYPE =
+  "azure_vm_uses_public_ip";
+export const SUBNET_VIRTUAL_MACHINE_RELATIONSHIP_TYPE = "azure_subnet_has_vm";
