+++ src/steps/resource-manager/key-vault/constants.ts
@@ -1,0 +1,14 @@
+import { generateRelationshipType } from "@jupiterone/integration-sdk";
+
+import { ACCOUNT_ENTITY_TYPE } from "../../active-directory";
+
+// Step IDs
+export const RM_KEYVAULT_VAULTS = "rm-keyvault-vaults";
+
+// Graph objects
+export const KEY_VAULT_SERVICE_ENTITY_TYPE = "azure_keyvault_service";
+export const ACCOUNT_KEY_VAULT_RELATIONSHIP_TYPE = generateRelationshipType(
+  "HAS",
+  ACCOUNT_ENTITY_TYPE,
+  KEY_VAULT_SERVICE_ENTITY_TYPE,
+);
