+++ src/steps/resource-manager/network/converters/securityGroups.ts
@@ -18,7 +18,7 @@ import {
   RelationshipDirection,
 } from "@jupiterone/integration-sdk";
 
-import { SECURITY_GROUP_RULE_RELATIONSHIP_TYPE } from "../../../../jupiterone";
+import { SECURITY_GROUP_RULE_RELATIONSHIP_TYPE } from "../constants";
 
 // interface SecurityGroupRelationship extends Relationship {
 //   _mapping?: RelationshipMapping;
