+++ src/steps/resource-manager/network/index.ts
@@ -7,9 +7,27 @@ import {
 import { Entity, Relationship } from "@jupiterone/integration-sdk";
 
 import { createAzureWebLinker } from "../../../azure";
-import { ACCOUNT_ENTITY_TYPE } from "../../../jupiterone";
 import { IntegrationStepContext } from "../../../types";
+import { ACCOUNT_ENTITY_TYPE, AD_ACCOUNT } from "../../active-directory";
 import { NetworkClient } from "./client";
+import {
+  LOAD_BALANCER_BACKEND_NIC_RELATIONSHIP_TYPE,
+  LOAD_BALANCER_ENTITY_TYPE,
+  NETWORK_INTERFACE_ENTITY_TYPE,
+  PUBLIC_IP_ADDRESS_ENTITY_TYPE,
+  RM_NETWORK_INTERFACES,
+  RM_NETWORK_LOAD_BALANCERS,
+  RM_NETWORK_PUBLIC_IP_ADDRESSES,
+  RM_NETWORK_SECURITY_GROUPS,
+  RM_NETWORK_VIRTUAL_NETWORKS,
+  SECURITY_GROUP_ENTITY_TYPE,
+  SECURITY_GROUP_NIC_RELATIONSHIP_TYPE,
+  SECURITY_GROUP_RULE_RELATIONSHIP_TYPE,
+  SECURITY_GROUP_SUBNET_RELATIONSHIP_TYPE,
+  SUBNET_ENTITY_TYPE,
+  VIRTUAL_NETWORK_ENTITY_TYPE,
+  VIRTUAL_NETWORK_SUBNET_RELATIONSHIP_TYPE,
+} from "./constants";
 import {
   createLoadBalancerBackendNicRelationship,
   createLoadBalancerEntity,
@@ -24,6 +42,8 @@ import {
   createVirtualNetworkSubnetRelationship,
   isWideOpen,
 } from "./converters";
+
+export * from "./constants";
 
 type SubnetSecurityGroupMap = {
   [subnetId: string]: NetworkSecurityGroup;
@@ -216,3 +236,53 @@ export async function fetchVirtualNetworks(
     await jobState.addEntity(createVirtualNetworkEntity(webLinker, vnet));
   });
 }
+
+export const networkSteps = [
+  {
+    id: RM_NETWORK_PUBLIC_IP_ADDRESSES,
+    name: "Public IP Addresses",
+    types: [PUBLIC_IP_ADDRESS_ENTITY_TYPE],
+    dependsOn: [AD_ACCOUNT],
+    executionHandler: fetchPublicIPAddresses,
+  },
+  {
+    id: RM_NETWORK_INTERFACES,
+    name: "Network Interfaces",
+    types: [NETWORK_INTERFACE_ENTITY_TYPE],
+    dependsOn: [AD_ACCOUNT, RM_NETWORK_PUBLIC_IP_ADDRESSES],
+    executionHandler: fetchNetworkInterfaces,
+  },
+  {
+    id: RM_NETWORK_VIRTUAL_NETWORKS,
+    name: "Virtual Networks",
+    types: [
+      VIRTUAL_NETWORK_ENTITY_TYPE,
+      SUBNET_ENTITY_TYPE,
+      VIRTUAL_NETWORK_SUBNET_RELATIONSHIP_TYPE,
+      SECURITY_GROUP_SUBNET_RELATIONSHIP_TYPE,
+    ],
+    dependsOn: [AD_ACCOUNT, RM_NETWORK_SECURITY_GROUPS],
+    executionHandler: fetchVirtualNetworks,
+  },
+  {
+    id: RM_NETWORK_SECURITY_GROUPS,
+    name: "Network Security Groups",
+    types: [
+      SECURITY_GROUP_ENTITY_TYPE,
+      SECURITY_GROUP_NIC_RELATIONSHIP_TYPE,
+      SECURITY_GROUP_RULE_RELATIONSHIP_TYPE,
+    ],
+    dependsOn: [AD_ACCOUNT, RM_NETWORK_INTERFACES],
+    executionHandler: fetchNetworkSecurityGroups,
+  },
+  {
+    id: RM_NETWORK_LOAD_BALANCERS,
+    name: "Load Balancers",
+    types: [
+      LOAD_BALANCER_ENTITY_TYPE,
+      LOAD_BALANCER_BACKEND_NIC_RELATIONSHIP_TYPE,
+    ],
+    dependsOn: [AD_ACCOUNT],
+    executionHandler: fetchLoadBalancers,
+  },
+];
