+++ src/steps/resource-manager/storage/constants.ts
@@ -1,0 +1,21 @@
+// Step IDs
+export const RM_STORAGE_RESOURCES = "rm-storage-resources";
+
+// Graph objects
+export const STORAGE_BLOB_SERVICE_ENTITY_TYPE = "azure_storage_blob_service";
+export const STORAGE_BLOB_SERVICE_ENTITY_CLASS = "Service";
+
+export const STORAGE_CONTAINER_ENTITY_TYPE = "azure_storage_container";
+export const STORAGE_CONTAINER_ENTITY_CLASS = "DataStore";
+
+export const STORAGE_FILE_SERVICE_ENTITY_TYPE = "azure_storage_file_service";
+export const STORAGE_FILE_SERVICE_ENTITY_CLASS = "Service";
+
+export const STORAGE_FILE_SHARE_ENTITY_TYPE = "azure_storage_share";
+export const STORAGE_FILE_SHARE_ENTITY_CLASS = "DataStore";
+
+export const STORAGE_QUEUE_SERVICE_ENTITY_TYPE = "azure_storage_queue_service";
+export const STORAGE_QUEUE_SERVICE_ENTITY_CLASS = "Service";
+
+export const STORAGE_TABLE_SERVICE_ENTITY_TYPE = "azure_storage_table_service";
+export const STORAGE_TABLE_SERVICE_ENTITY_CLASS = "Service";
